Improve performance of updates to ax_info and other misc improvements

+++ src/utils_gen/structures_utils.jl
@@ -15,12 +15,12 @@ function installed()
 end
 
 """
-    place_nodes_in_circle(total_nodes::Int, current_node::Int, r::Float32, xₒ::Float32, yₒ::Float32)
+    place_nodes_in_circle(total_nodes::Int64, current_node::Int64, r::Float32, xₒ::Float32, yₒ::Float32)
 
 Return coordinate for point number `i` of a total of `n` points evenly distributed around
 a circle of radius `r` centered at (xₒ, yₒ) from -π/4 to 5π/4.
 """
-function place_nodes_in_circle(n::Int, i::Int, r::Float32, xₒ::Float32, yₒ::Float32)
+function place_nodes_in_circle(n::Int64, i::Int64, r::Float32, xₒ::Float32, yₒ::Float32)
     θ::Float32 = n == 1 ? π : -π / 4 + 3π / 2 * (1 - (i - 1) / (n - 1))
     x::Float32 = xₒ + r * cos(θ)
     y::Float32 = yₒ + r * sin(θ)
